GUI stuff added
I'm so sorry I deleted Main, FML.
